Cleanup Brave Ads ad event utils
