+++ components/brave_ads/core/internal/BUILD.gn
@@ -190,12 +190,12 @@ source_set("internal") {
     "ad_constants.cc",
     "ad_type.cc",
     "ads.cc",
+    "ads/ad_events/ad_event_handler_util.cc",
+    "ads/ad_events/ad_event_handler_util.h",
     "ads/ad_events/ad_event_history.cc",
     "ads/ad_events/ad_event_info.cc",
     "ads/ad_events/ad_event_info.h",
     "ads/ad_events/ad_event_interface.h",
-    "ads/ad_events/ad_event_util.cc",
-    "ads/ad_events/ad_event_util.h",
     "ads/ad_events/ad_events.cc",
     "ads/ad_events/ad_events.h",
     "ads/ad_events/ad_events_database_table.cc",
@@ -299,7 +299,11 @@ source_set("internal") {
     "ads/serving/eligible_ads/allocation/round_robin_ads.h",
     "ads/serving/eligible_ads/allocation/round_robin_advertisers.h",
     "ads/serving/eligible_ads/allocation/seen_ads.h",
+    "ads/serving/eligible_ads/allocation/seen_ads_util.cc",
+    "ads/serving/eligible_ads/allocation/seen_ads_util.h",
     "ads/serving/eligible_ads/allocation/seen_advertisers.h",
+    "ads/serving/eligible_ads/allocation/seen_advertisers_util.cc",
+    "ads/serving/eligible_ads/allocation/seen_advertisers_util.h",
     "ads/serving/eligible_ads/eligible_ads_alias.h",
     "ads/serving/eligible_ads/eligible_ads_callback.h",
     "ads/serving/eligible_ads/eligible_ads_constants.h",
