+++ components/brave_ads/core/internal/ads/serving/eligible_ads/allocation/seen_ads_util.h
@@ -1,10 +1,10 @@
-/* Copyright (c) 2021 The Brave Authors. All rights reserved.
+/* Copyright (c) 2023 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_AD_EVENTS_AD_EVENT_UTIL_H_
-#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_AD_EVENTS_AD_EVENT_UTIL_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_SERVING_ELIGIBLE_ADS_ALLOCATION_SEEN_ADS_UTIL_H_
+#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_SERVING_ELIGIBLE_ADS_ALLOCATION_SEEN_ADS_UTIL_H_
 
 #include "brave/components/brave_ads/core/internal/ads/ad_events/ad_event_info.h"
 #include "third_party/abseil-cpp/absl/types/optional.h"
@@ -15,20 +15,11 @@ class Time;
 
 namespace brave_ads {
 
-class ConfirmationType;
-struct AdInfo;
 struct CreativeAdInfo;
-
-bool HasFiredAdEvent(const AdInfo& ad,
-                     const AdEventList& ad_events,
-                     const ConfirmationType& confirmation_type);
 
 absl::optional<base::Time> GetLastSeenAdTime(const AdEventList& ad_events,
                                              const CreativeAdInfo& creative_ad);
-absl::optional<base::Time> GetLastSeenAdvertiserTime(
-    const AdEventList& ad_events,
-    const CreativeAdInfo& creative_ad);
 
 }  // namespace brave_ads
 
-#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_AD_EVENTS_AD_EVENT_UTIL_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ADS_SERVING_ELIGIBLE_ADS_ALLOCATION_SEEN_ADS_UTIL_H_
