+++ components/brave_ads/core/test/BUILD.gn
@@ -101,9 +101,9 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/account/wallet/wallet_unittest_constants.h",
     "//brave/components/brave_ads/core/internal/account/wallet/wallet_unittest_util.cc",
     "//brave/components/brave_ads/core/internal/account/wallet/wallet_unittest_util.h",
+    "//brave/components/brave_ads/core/internal/ads/ad_events/ad_event_handler_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/ad_events/ad_event_history_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/ad_events/ad_event_unittest_util.cc",
-    "//brave/components/brave_ads/core/internal/ads/ad_events/ad_event_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/ad_events/ad_events_database_table_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/ad_events/inline_content_ads/inline_content_ad_event_handler_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/ad_events/new_tab_page_ads/new_tab_page_ad_event_handler_if_ads_disabled_unittest.cc",
@@ -137,6 +137,8 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/ads/search_result_ad_test.cc",
     "//brave/components/brave_ads/core/internal/ads/serving/choose/eligible_ads_predictor_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/serving/choose/sample_ads_unittest.cc",
+    "//brave/components/brave_ads/core/internal/ads/serving/eligible_ads/allocation/seen_ads_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/ads/serving/eligible_ads/allocation/seen_advertisers_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/serving/eligible_ads/eligible_ads_feature_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/serving/eligible_ads/eligible_ads_feature_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/ads/serving/eligible_ads/eligible_ads_unittest_util.cc",
